backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
python-multipart
loguru
